+++ src/main/java/com/seyed/ali/timeentryservice/service/interfaces/TimeEntryService.java
@@ -12,13 +12,8 @@ public interface TimeEntryService {
 
     String addTimeEntryManually(TimeEntryDTO timeEntryDTODTO);
 
-    // TODO: Implement REDIS for getting the cached `start_time`
-    TimeEntryDTO stopTrackingTimeEntry(String timeEntryId);
-
     TimeEntryDTO updateTimeEntryManually(String id, TimeEntryDTO timeEntryDTO);
 
     void deleteTimeEntry(String id);
-
-    String startTrackingTimeEntry();
 
 }
