+++ src/test/java/com/seyed/ali/timeentryservice/service/TimeEntryServiceImplTest.java
@@ -159,84 +159,6 @@ class TimeEntryServiceImplTest extends TimeParserUtilForTests {
     }
 
     @Test
-    void startTrackingTimeEntryTest() {
-        // given
-        when(this.authenticationServiceClient.getCurrentLoggedInUsersId())
-                .thenReturn("some_user_id");
-        when(this.timeEntryRepository.save(isA(TimeEntry.class)))
-                .thenReturn(this.timeEntry);
-
-        // when
-        String timeEntryId = this.timeEntryService.startTrackingTimeEntry();
-
-        // then
-        assertThat(timeEntryId)
-                .isNotNull();
-
-        verify(this.timeEntryRepository, times(1))
-                .save(isA(TimeEntry.class));
-    }
-
-    /*@Test // TODO: implement testing for this - it Fu**d me up :(
-    public void stopTrackingTimeEntryTest() {
-        // Arrange
-        String userId = "testUserId";
-        LocalDateTime startTime = LocalDateTime.of(2024, 5, 11, 8, 0, 0);
-        LocalDateTime endTime = LocalDateTime.of(2024, 5, 11, 10, 0, 0);
-
-        TimeEntry timeEntry = new TimeEntry();
-        timeEntry.setId("1");
-        timeEntry.setStartTime(this.startTime);
-        timeEntry.setUserId(userId);
-
-        when(this.authenticationServiceClient.getCurrentLoggedInUsersId())
-                .thenReturn(userId);
-
-        when(this.timeEntryRepository.findByUserIdAndId(isA(String.class), isA(String.class)))
-                .thenReturn(timeEntry);
-
-        when(this.timeEntryRepository.save(isA(TimeEntry.class)))
-                .thenReturn(timeEntry);
-
-        when(this.timeParser.parseLocalDateTimeToString(this.startTime))
-                .thenReturn(this.startTimeStr);
-
-        LocalDateTime fixedEndTime = this.endTime;
-        MockedStatic<LocalDateTime> localDateTimeMockedStatic = mockStatic(LocalDateTime.class);
-        localDateTimeMockedStatic
-                .when(LocalDateTime::now)
-                .thenReturn(fixedEndTime);
-        LocalDateTime endTimeNow = LocalDateTime.now();
-        String endTimeNowStr = this.parseLocalDateTimeToString(fixedEndTime);
-        when(this.timeParser.parseLocalDateTimeToString(endTimeNow))
-                .thenReturn(endTimeNowStr);
-
-        when(this.timeParser.parseDurationToString(isA(Duration.class)))
-                .thenReturn(this.durationStr);
-
-        // Act
-        TimeEntryDTO result = this.timeEntryService.stopTrackingTimeEntry(timeEntry.getId());
-        System.out.println(result);
-
-        // Assert
-        assertThat(result)
-                .isNotNull();
-        assertThat(result.startTime())
-                .as("StartTimeString must be same")
-                .isEqualTo(this.startTimeStr);
-        assertThat(result.endTime())
-                .as("EndTimeString must be same")
-                .isEqualTo(this.endTimeStr);
-        assertThat(result.duration())
-                .as("DurationString must be same")
-                .isEqualTo(this.durationStr);
-
-        verify(this.timeEntryRepository).save(timeEntry);
-
-        localDateTimeMockedStatic.close();
-    }*/
-
-    @Test
     public void updateTimeEntryTest() {
         // Given
         String id = "Some_timeEntry_id";
